Resolves #66 and improves terminal scrolling buffer issues.
1. Allowed the logic to allow having lines in the TextEditor subbuffer
causing the scrolling of the underlying ScreenBuffer when the text
overuns the bottom of the screen.

2. Now puts the cursor in the right place and aligns the screen
correctly when using up-arrow to obtain a previously entered long
line that wrapped past the edge of the screen.

3. Fixed the off-by-one error where typing a line that wrapped around
past the right edge of the screen used to put the cursor in the
wrong place.  The more times the line wrapped the more off the cursor
became.