First checkin of textedit popup window.  Still a bit rough around the edges.
NOTE: TO get this to work you also need to copy a PNG image file and install
it.  I think there needs to be a location in the git archive for the auxilliary
files that are in the installable ZIP but are not C# code (like the model mesh,
texture images, and so on)  But since there isn't one yet I didn't put my
image inside the git archive yet.

Get my image (which I will post onto github with this commit message) and
Save it as:

GameData/kOS/GFX/resize-button.gif