TextEditor ready for a release.  Everything works but the scrollbar.
The issue with the scrollbar is caused by a problem I know, but it would
take a bit of effort to fix it.  The scrollbar can be dragged with the
mouse, but it will only let you scroll within the range of the text cursor,
not allowing the viewport to move far enough to put the text cursor offscreen.

Fixed all the following though:

* There is now a confirmation dialog when something will throw away unsaved changes.  Dirty flag check implemented.
* Resize corner works.  The window isn't drawn during the resize, but that's okay.
* Title shows which volume the file is on (but it can't show the volume number of an unnamed volume because that's not contained in the Volume object - it's in the VolumeManager and I avoided adding the _shared global to this object so I can't access the VolumeManager.
* TextArea is now a GuiLayout.  That means it gets a scrollbar indicator.
* Added Keypresses: PgUp, PgDown, Control + S, E, or R for save/exit/reload.
* Removed triple-save problem.
* Better messages on terminal.
* When file save fails (ie. Volume is full), it lets you cancel and keep the window up.

NOTE: This *will* expose to the users the fact that the local volume has been
losing line breaks in savegames this whole time and nobody knew because you
could never look at files on the local volume before.  That's not a bug caused
by this change.  It's a bug that was always there that wasn't noticed before.