Added the code to implement the following user-callable functions:
POSITIONAT( THING, TIME )
Returns a prediction of where ''thing'' will be at universal timestamp ''time''.
If the THING is a Vessel, and the Vessel has planned manuever nodes, the
prediction assumes the manuever nodes will be executed exactly as planned.
  ''thing'' is an Orbitable
  ''time'' is a Timestamp
  ''return value'' is a VECTOR, expressed in ship-center-raw-rotation coords.

VELOCITYAT( THING, TIME )
Returns a prediction of what ''thing''s velocity will be at universal timestamp ''time''.
If the THING is a Vessel, and the Vessel has planned manuever nodes, the
prediction assumes the manuever nodes will be executed exactly as planned.
  ''thing'' is an Orbitable
  ''time'' is a Timestamp
  ''return value'' is an ORBITABLEVELOCITY structure.

ORBITAT( THING, TIME )
Returns the Orbit patch where the thing is predicted to be at universal timestamp time.
If the THING is a Vessel, and the Vessel has planned manuever nodes, the
prediction assumes the manuever nodes will be executed exactly as planned.
  ''thing'' is an Orbitable
  ''time'' is a Timestamp
  ''return value'' is an ORBIT structure.