Resolves #196
Apparently the persistence save file itself also supports comments
which use the exact same double-slash syntax.  So when the kOS file
contents were being stored in the savegame as a single encoded long
string in the persistence file, the "//" within the value was commenting
out the rest of the value from KSP's own savegame parser.