Resolves #207 Mod compatibility (Module Manager(2.2.0))
This fixes null object references which happen when you use Module
Manager to add the kOSProcessor module to command parts which are the
root part of the ship.  These errors will happen every time you Revert,
load either from Persistent or Quicksave, change ships, or change
scenes.

In FlightControlManager.cs:
AddTo is rearranged so that it only tries to work when both
shared.Vessel and shared.Vessel.rootPart are both not null.
UnBind won't try to get and unbind the flightControls entry if
currentVessel.rootPart is null.
Dispose won't try to UnBind or Remove the flight control when
currentVessel or currentVessel.rootPart are null.