There was a bug with the Close button on the terminal only working part of the time, which I traced to the fact that the PNG image file for the terminal hardcoded the Close button into the terminal image, which covered up the actual close button that the GUI was trying to make. Before this hadn't caused a problem because we were allowing clicks to pass through all layers.  But when I made clicks stop at the first object they hit, that meant the terminal image of a fake "close" button was blocking the clicks from reaching the actual close button underneath.
I removed the close button from the terminal image,
so it always gets created dynamically by the GUI window
paint routine instead, and altered the buttons of the
popup editor to always use the Unity GUI.Button type
instead of trying to roll-my-own button out of a box.
this is much cleaner with the skinning and the window
stacking.