More exceptions for issue #254.
There seems to be a bug where line number information is missing
and this causes the compiler to claim an error is internal when
it's really user-caused.  Not sure where that is, but I'm not entirely
convinced I caused it because I remember getting it before this change.
Fixing it should be handled as a separate issue after this is merged in.