Resolves #257.  (NOTE THIS USES MASTER BRANCH, not Develop)
This edit was based from Master and not Develop specifically because
there's good rumors that KSP 0.25 may be out in a week, and this
issue might need a quickfix put out before the heavy changes we're
doing in develop are ready for release.

Please note the "TODO for KSP 0.25" notices in the comments, and
be ready to quickly implement what they're talking about when 0.25
is out.