method call return null no longer gives bogus error.
Previously, if you had this:
    // FOO IS some sort of list.
    SET ITER TO FOO:ITER.

And did this:
    ITER:RESET.

Instead of this:
    ITER:RESET().

Then it *would* work, but after calling Enumerator.GetMethod("RESET")
it would incorrectly claim "RESET" is not a suffix of ITER.  This was
because Structure.GetMethod() uses null returns as its way of
communicating that a suffix doesn't exist.  It should have used an
exception, which is what it does now.