Added PARTSINGROUP and MODULESINGROUP to Vessel.
Got these methods working now.

NOTE it's an annoyingly nested algorithm because KSP
doesn't maintain a link from action groups to parts.
(or at least if it does I don't know where it is in the API.)

I have to look at all the parts and partmodules and see
which of them have the group I'm looking for on them.