suffix and array change fix part 1.
This commit fixes the following parsing problems

```
aaa:bbb[ccc]
aaa:bbb[ccc]:ddd(eee)
```

It still does not fix this problem:

```
aaa:bbb(ccc)[ddd]
```

An example of that problem in practice:

```
set x to ship:partsnamed("rtg")[0].  // the zero-th RTG part on the ship.
```

Fixing this problem means editing the grammar file, as it just doesn't
even allow that formation at all the way it's defined now.  That leaves
it as a problem for tomorrow.  Tonight I'm sleepy and want to get this
committed before I quit for the night.