Resolves #318 and adds a few related fixes.
Other things dealt with are:

:MODULES suffix now has an alias :ALLMODULES that does the same
thing.  (Reason: the exception message for :GETMODULE("some bogus name")
tells the user to try running the :ALLMODULEs suffix, because it
always inserts the word "ALL" in front of the argument passed into
it to form the message.  I figured it might be nice if the thing
the excpetion tells the users to do is in fact actually possible.)

Restructured the "is this thing visible" checks to collect them all
into one function.

Took into account whether the editor is running or the flight mode
is running, to change whether it uses guiActive or guiActiveEditor.
At the moment this doesn't matter much because you can't run programs
inside the editor... but some day.....