Fiddling with GUI layouts on the app panel.
Since @Erendrake appears to have gone dark for a bit, and I assume that's
baby related, I went ahead and took the time to pretty-up the kOS popup
panel, moving the long explanatory text for the options to a tooltip
at the bottom of the panel so it takes less space, and assigning a
layout pattern to use for the future things on the panel (terminal
selector, part nametag selector).  Also, finally worked out how to
make a dynamically altering window size that just magically fits
the content without specifying fixed pixel size ahead of time.

I still can't figure out how to get the toggle buttons to stop having
so much padding space above and below them.  I can't figure out where
that padding is coming from.  I played with the margin, border, and padding
settings and they didn't affect it so I gave up and left them a bit big.
In the long run it won't matter because the other content of the window
will keep it that big anyway, probably.

Adjunct to this, the following related changes were made in service of
it:

New Utils function to get the current version number directly from
the DLL assembly just in case you want to ever display it. (I used
it as part of the title of the app panel.)

CONFIG now supports the ability to have a max and min limitation
for a value, which I used to constrain CONFIG:IPU to the range
between 50 and 2000 inclusive.  If you try to set it to a value
outside this range, it silently clamps it down to within this
range without raising a complaint.  (If there were such a thing
as a KOS Warning rather than Error, this would be a good place for
one).  This was done so that users typing randon garbage into the
field on the app panel won't break the system.  However, it also
works for in-code things, like saying SET CONFIG:IPU TO 1000000
will result it it only becoming 2000.