Resolves #379.
The actual problem, after looking at it, only happened
on the error where you forgot the closing dot of a command,
like you did this:

    print "hello"

instead of this

    print "hello.

It didn't happen on other syntax errors, and it only happened
in the live interpreter.

The cause was the fact that the interpreter appends an
OpcodeEOF to the end of the program context in a way that
is outside the normal compile process, and in so doing
it's missing the debug data, like so:

  Code Fragment
  File                 Line:Col IP   opcode operand
  ----                 ----:--- ---- ---------------------
			  0:0   0000 jump 1
  interpreter history     2:1   0001 push $sfasdfsd
  interpreter history     2:1   0002 pop
  interpreter history     4:7   0003 push hello
  interpreter history     4:7   0004 call print()
  interpreter history     5:7   0005 push $asdf
  interpreter history     5:7   0006 call print()
                          0:0   0007 EOF  <<--INSTRUCTION POINTER

So when the syntax error happens at the EOF, it doesn't
know during the stack trace that it came from user typing.
The lack of a program name or "interpreter history" makes it
believe it came from an internal problem.

The solution I implemented was to perform a special case in KSPLogger
when the opcode is OpcodeEOF and bypass the normal attempts
to pull it out of the source line - there is no source line,
its after the end of the source code that the error happened.