Last minute bug cleanup on file extensions.
This is the result of my checking the Volume vs Archive
behavior, and discovering a few weird edge cases where
running a sub-program from a program behaved differently
from running a program from command-line.

Also tested that the logic of compiled vs source running
is all good, and made a few tweaks to get that working.