Skin copy doesn't work right.  This is what I feared.
The TextArea for Config:IPU threw exceptions whenever
you tried typing into it.  It turns out to be because
of the trickery I tried to get around the lack of a
copy operator for GUISkin.  It seems to have left
something behind uncopied that was still null when
the renderer tried using it.

The lack of a proper guiskin copy meant one aspect of the
skin didn't copy and I don't know what it is.  Instead
of trying to find it I'm not using the skin entirely
and instead just using the SQUAD default and then overriding
the style of every widget I draw, to get sort of the same
effect.  Later I can re-visit what to do to make copying
a gui skin work right.