Syntax fixes for arrays Fixes #387
When the syntax tree had new intermediate parse nodes
added to it to handle better genericicity of what
can come in what order (suffixes of arrays, arrays of
suffixes, suffixes that are method calls, etc), it
seems to have broken two areas where the code explicitly
had to walk the parse tree to answer questions like
"am I the lastmost thing in this chain of suffixes?"
or "Am I an array index at the end of the chain, or the middle?"

Those tree walks failed to take into account the new nesting
levels and the new syntax nodes the tree walk would end
up encountering.

It should be fixed now.