Fix for filenames without extensions
Added a call to "CookedFilename" when initializing the FileName property
on the ProgramFile class.  This should prevent us from accidentally
creating any files without an extension, while still allowing users to
set their own extension when referring to files.  It appears to work
with all of my existing scripts on vessels, and fixes all of my boot
scripts.