Removed dir:FROM, added dir:INVERSE instead.
The :FROM suffix wasn't working right at all.  Instead I decided
to offer up the Unity "Inverse" quaternion operator, which can be
used to obtain a "FROM" way of thinking with a few more steps.

Normally a Direction in kOS is a rotation that goes from the native
XYZ axes TO a new reference frame in which the Z-axis is pointing
the way you want to go.

What the :INVERSE suffix of a direction will get you is the rotation
the other way around - the rotation that would go from the Direction
TO the native XYZ axes.  Its used for getting the attributes of a
vector converted into the ref frame relative to the direction.