Fixes #447
The problem is that I was using the states of the game as a bitfield, when
they aren't in fact set up as a bitfield (the values don't go 1,2,4,8,16, etc).

I was trying to set up the applauncher to trigger on EditorAny|Flight, but
you can't bit-OR them like that to get that effect.  SQUAD didn't design them
that way.

So now it just activates in ALL scenes, since it only has to run the setup
once, and then after that it will suppress its own Start() behavior on all
additional activations.