Workaround to fix #448 and maybe #428
RemoteTechArchive.cs - Modified CheckRange to return true if the vessel
is currently in the prelaunch condition.  This prevents the archive from
showing as out of comm range during initial start up when RemoteTech
seems to have issues returning the correct value reliably.
BindingManager.cs - Added a null check before disposing the
flightControl variable.  I have had issues on and off with this throwing
null reference errors when reverting or going to the space center.